pyproject.toml
[tool.poetry]
name = "merenda"
version = "1.0"
description = "Descrição sobre o projeto"
authors = ["Valtair Junior <valtair.junior@unemat.br>"]
readme = "README.md"
packages = [{include = "merenda"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.5.post1"
uvicorn = "^0.20.0"
python-magic = "^0.4.27"
pyjwt = "^2.6.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
psycopg2-binary = "^2.9.5"
alembic = "^1.10.2"
pydantic-settings = "^2.0.3"
fastapi = "^0.104.0"
email-validator = "^2.1.0.post1"
mkdocstrings = "^0.24.0"
griffe = "^0.40.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.1.0"
taskipy = "^1.10.3"
pre-commit = "^3.1.1"


[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.21.0"
sqlalchemy-utils = "^0.40.0"
faker = "^18.3.1"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
httpx = "^0.23.3"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.5.8"
mkdocstrings-python = "^1.8.0"
mkdocs-swagger-ui-tag = "^0.6.8"
jinja2 = "^3.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules --ignore=alembic"

[tool.black]
line-length = 79

[tool.isort]
line_length = 79
profile = "black"

[tool.taskipy.tasks]
run = "uvicorn app:app --reload --host 0.0.0.0 --port 7000 --proxy-headers --forwarded-allow-ips '*'"
lint = "black --check --diff . && isort --check --diff ."
docs = "mkdocs serve"
test = "pytest -s -x --cov=app -vv"
test_pdb = "pytest -s -x --pdb"
post_test = "coverage html"
cov = "google-chrome htmlcov/index.html"
rev = "read -p 'nome da revision: ' nome && alembic revision --autogenerate -m $nome "
up = "alembic upgrade head"
down = "alembic downgrade base"
export = "poetry export > requirements.txt --without-hashes"
install_req = "cat requirements.txt | grep -E '^[^# ]' | cut -d ';' -f1  | xargs -n 1 poetry add"
install_up_req = "cat requirements.txt | grep -E '^[^# ]' | cut -d '=' -f1  | xargs -n 1 poetry add"
pc = "pre-commit run --all-files"
pc_update = "pre-commit autoupdate"
init_db = "docker start postgres"
